build(cucascade): build against PR #150 cudf-free-core split (#989)
cuCascade PR #150 splits the library into a cudf-free core
(cuCascade::cucascade) and an optional cuCascade::cucascade_cudf target
holding the cudf-coupled representations, converters, and bandwidth
profiler. Point the submodule at the PR head and adapt Sirius to the new
layout.

CMake:
- Force CUCASCADE_BUILD_CUDF=ON and link cuCascade::cucascade_cudf into
both extension targets and parquet_benchmark (it transitively pulls the
core + cudf::cudf).
- Add the cucascade_cudf_* targets to the VCPKG static-cudart patching
loop and cucascade_cudf_static to the DuckDB install/export set.

Includes (headers moved data/,memory/ -> cudf/):
- data/cpu_data_representation.hpp -> cudf/host_data_representation.hpp
- data/gpu_data_representation.hpp -> cudf/gpu_data_representation.hpp
- memory/host_table.hpp           -> cudf/host_table.hpp
- Include cudf/builtin_converters.hpp where
register_builtin_converters() is
  called (it moved out of representation_converter.hpp).
- Add explicit cudf/* includes to files that previously obtained these
symbols
  transitively via data_batch.hpp, which is now cudf-free.

column_metadata: its type_id field changed from cudf::type_id to an
opaque int32_t tag. Cast at write sites (static_cast<int32_t>) and read
sites (static_cast<cudf::type_id>), matching cucascade's as_cudf_type_id
idiom.

Builds clean (extension + unittest); the converter/host-table/scan unit
tests pass on RTX PRO 6000.

---------

+++ src/op/scan/duckdb_scan_task.cpp
@@ -29,7 +29,7 @@
 #include <op/scan/duckdb_scan_task.hpp>
 
 // cucascade
-#include <cucascade/data/cpu_data_representation.hpp>
+#include <cucascade/cudf/host_data_representation.hpp>
 #include <cucascade/memory/memory_reservation.hpp>
 
 // duckdb
@@ -265,7 +265,7 @@ duckdb_scan_task_local_state::column_builder::make_column_metadata(size_t num_ro
   if (type.is_varchar()) {
     // VARCHAR column: data buffer + offsets child
     column_metadata offsets_child{};
-    offsets_child.type_id          = cudf::type_id::INT32;
+    offsets_child.type_id          = static_cast<int32_t>(cudf::type_id::INT32);
     offsets_child.num_rows         = static_cast<cudf::size_type>(num_rows + 1);
     offsets_child.null_count       = 0;
     offsets_child.scale            = 0;
@@ -277,7 +277,7 @@ duckdb_scan_task_local_state::column_builder::make_column_metadata(size_t num_ro
     offsets_child.data_size        = (num_rows + 1) * sizeof(int32_t);
 
     column_metadata col{};
-    col.type_id          = cudf::type_id::STRING;
+    col.type_id          = static_cast<int32_t>(cudf::type_id::STRING);
     col.num_rows         = static_cast<cudf::size_type>(num_rows);
     col.null_count       = static_cast<cudf::size_type>(null_count);
     col.scale            = 0;
@@ -294,7 +294,7 @@ duckdb_scan_task_local_state::column_builder::make_column_metadata(size_t num_ro
     auto cudf_type = sirius::get_cudf_type(type);
 
     column_metadata col{};
-    col.type_id          = cudf_type.id();
+    col.type_id          = static_cast<int32_t>(cudf_type.id());
     col.num_rows         = static_cast<cudf::size_type>(num_rows);
     col.null_count       = static_cast<cudf::size_type>(null_count);
     col.scale            = cudf_type.scale();
